+++ index.html
@@ -7,8 +7,8 @@
   <link rel="stylesheet" href="css/normalize.css">
   <link rel="stylesheet" href="css/style.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
+  <script src="js/main.js" type="module" defer></script>
   <script src="js/function.js" defer></script>
-  <script src="js/main.js" defer></script>
   <title>Кекстаграм</title>
 </head>
